+++ examples/cc_lb/variables.tf
@@ -213,6 +213,12 @@ variable "load_distribution" {
   }
 }
 
+variable "encryption_at_host_enabled" {
+  type        = bool
+  description = "User input for enabling or disabling host encryption"
+  default     = false
+}
+
 
 # Azure Private DNS specific variables
 variable "zpa_enabled" {